requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Pillow==10.2.0
reportlab==4.1.0
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Cors==4.0.0
gunicorn==21.2.0
qrcode==7.4.2
